+++ lib/twofishes/client.rb
@@ -11,10 +11,10 @@ module Twofishes
     # @example
     #   Twofishes::Client.geocode('Zurich, Switzerland')
     #
-    def self.geocode(location, includes = [])
+    def self.geocode(location, includes: [])
       handle_response do
         request = GeocodeRequest.new(query: location, responseIncludes: includes)
-        thrift_client.geocode()
+        thrift_client.geocode(request)
       end
     end
 
@@ -25,7 +25,7 @@ module Twofishes
     # @example
     #   Twofishes::Client.reverse_geocode([47.3787733, 8.5273363])
     #
-    def self.reverse_geocode(coordinates, includes = [])
+    def self.reverse_geocode(coordinates, includes: [])
       handle_response do
         point = GeocodePoint.new(lat: coordinates[0], lng: coordinates[1])
         request = GeocodeRequest.new(ll: point, responseIncludes: includes)