+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Now uses thrift instead of REST. #4
 - Up to [60% faster](https://github.com/masone/twofishes-ruby/pull/4#issuecomment-74030852).
 - `Twofishes.configure` now accepts `host` and `port` respectively. Be aware you need to provide thrift port (8080 by default), not http server port (8081 by default).
-- The response no longer includes `parents` by default. You can easily get them by passing `[ResponseIncludes::PARENTS]` as the second parameter to `geocode`. Look at the README for an example.
+- The response no longer includes `parents` by default. You can easily get them by passing `[ResponseIncludes::PARENTS]` as the `:includes` keyword argument to `geocode`. Look at the README for an example.
 
 ## [1.0.1] - 2014-12-08